Move the official/benchmark files to offical/utils/logging.
